File moved to Cryptography -> crc32 .h/cpp
File moved to Cryptography -> LogonCommDefines.h
