+++ src/shared/Cryptography/LogonCommDefines.h
