+++ src/shared/Cryptography/crc32.h
